test(domain): split usage tests into test_usage_models.py
Mirrors the production split between qfa.domain.models and
qfa.domain.usage_models. Moved all usage-tracking and aggregation
tests to the new test_usage_models.py file, keeping non-usage tests
(requests, responses, API keys, sensitivity) in test_models.py.
